requirements.txt
Flask==3.0.0
flask-cors==4.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers==4.35.0
Pillow>=10.0.0
numpy==1.24.3
requests==2.31.0
